session/src/main/java/by/innowisegroup/session/session/MySessionFactory.java
package by.innowisegroup.session.session;

import org.hibernate.SessionFactory;

import org.hibernate.cfg.Configuration;

public class MySessionFactory {

    private static SessionFactory sessionFactory = buildSessionFactory();

    private static SessionFactory buildSessionFactory() {
        try {
            if (sessionFactory == null) {
                Configuration config = new Configuration();
                config.configure();
                sessionFactory = config.buildSessionFactory();
            }
            return sessionFactory;
        } catch (Throwable ex) {
            throw new ExceptionInInitializerError(ex);
        }
    }

    public static SessionFactory getSessionFactory() {
        return sessionFactory;
    }

    public static void shutdown() {
        getSessionFactory().close();
    }
}